requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Django==4.1.7
djangorestframework==3.14.0
dnspython==2.3.0
Flask==2.2.2
heyoo==0.0.8
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mongoengine==0.27.0
pymongo==4.3.3
python-decouple==3.8
pytz==2022.7.1
requests==2.28.1
requests-toolbelt==0.9.1
sqlparse==0.4.3
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
